refactor: remove usePrice hook
